finish Tests

+++ src/main/java/by/piskunou/solvdlaba/persistence/FlightRepository.java
@@ -17,7 +17,7 @@ public interface FlightRepository {
     Optional<Flight> findById(long id);
 
     List<Flight> search(@Param("fromAirports") List<String> fromAirports, @Param("toAirports") List<String> toAirports,
-                        @Param("passengerAmount") long passengerAmount,
+                        @Param("passengerAmount") int passengerAmount,
                         @Param("start") LocalDateTime start, @Param("end") LocalDateTime end);
 
     List<Seat> flightSeats(long id);
